refactor : [ 위치 변경 및 필요없는 코드 수정 ]
